requirements.txt
Flask==2.3.2
flask-cors==4.0.0
pymongo==4.4.0
python-dotenv==1.0.0
gunicorn==21.2.0
